+++ x/fungible/types/keys.go
@@ -29,7 +29,10 @@ func KeyPrefix(p string) []byte {
 var (
 	ModuleAddress    = authtypes.NewModuleAddress(ModuleName)
 	ModuleAddressEVM = common.BytesToAddress(ModuleAddress.Bytes())
-	AdminAddress     = "zeta1rx9r8hff0adaqhr5tuadkzj4e7ns2ntg446vtt"
+	// ModuleAddressZEVM is calculated in the same way as ModuleAddressEVM.
+	// Maintain it for legibility in functions calling fungible module address from zEVM.
+	ModuleAddressZEVM = common.BytesToAddress(ModuleAddress.Bytes())
+	AdminAddress      = "zeta1rx9r8hff0adaqhr5tuadkzj4e7ns2ntg446vtt"
 )
 
 const (